+++ examples/cube3d/src/main.rs
@@ -13,19 +13,8 @@ use bevy::{
 };
 use bevy_vello::{VelloPlugin, prelude::*, render::VelloRenderer};
 
-#[derive(Component)]
+#[derive(Component, Clone, ExtractComponent)]
 pub struct VelloTarget(Handle<Image>);
-
-impl ExtractComponent for VelloTarget {
-    type QueryData = &'static VelloTarget;
-    type QueryFilter = ();
-    type Out = Self;
-    fn extract_component(
-        target: bevy::ecs::query::QueryItem<'_, '_, Self::QueryData>,
-    ) -> Option<Self> {
-        Some(Self(target.0.clone()))
-    }
-}
 
 // Marks the main pass cube, to which the texture is applied.
 #[derive(Component)]