+++ src/render/extract.rs
@@ -1,7 +1,10 @@
 use bevy::{
     prelude::*,
     render::{MainWorld, extract_component::ExtractComponent},
+    window::PrimaryWindow,
 };
+
+use crate::render::VelloPixelScale;
 
 use super::{VelloEntityCountData, VelloFrameProfileData};
 
@@ -26,19 +29,13 @@ pub fn sync_frame_profile(render_data: Res<VelloFrameProfileData>, mut world: Re
 }
 
 /// A screenspace render target. We use a resizable fullscreen quad.
-#[derive(Component, Default)]
+#[derive(Component, Default, Clone, ExtractComponent)]
 pub struct SSRenderTarget(pub Handle<Image>);
 
-impl ExtractComponent for SSRenderTarget {
-    type QueryData = &'static SSRenderTarget;
-
-    type QueryFilter = ();
-
-    type Out = Self;
-
-    fn extract_component(
-        ss_render_target: bevy::ecs::query::QueryItem<'_, '_, Self::QueryData>,
-    ) -> Option<Self> {
-        Some(Self(ss_render_target.0.clone()))
-    }
+pub fn extract_pixel_scale(
+    mut pixel_scale: ResMut<VelloPixelScale>,
+    window: Single<&Window, With<PrimaryWindow>>,
+) {
+    let scale_factor = window.resolution.scale_factor();
+    pixel_scale.0 = scale_factor;
 }